complete: ownership confirmation of note before update and delete

+++ src/controllers/notes.controllers.ts
@@ -39,15 +39,21 @@ export const getNotes = async (req: Request, res: Response) => {
 
 export const updateNote = async (req: Request, res: Response) => {
     try {
-        const { notesId } = req.params;
+        const { noteId } = req.params;
         const { title, description, tags } = req.body;
         if (!title || !description) {
             return res.status(400).json({ "message": "Incomplete Information" })
         }
-
-        const note = await GetNoteById(notesId);
+        const note = await GetNoteById(noteId);
         if (!note) {
             return res.status(400).json({ "message": "Note Id not matched" });
+        }
+        const currentUserId = get(req, 'identity._id') as string | undefined;
+        if (!currentUserId) {
+            return res.sendStatus(403);
+        };
+        if (currentUserId.toString() !== note.userId) {
+            return res.status(400).json({ "message": "Ownership mismatch" })
         }
         note.title = title;
         note.description = description;
@@ -64,14 +70,21 @@ export const updateNote = async (req: Request, res: Response) => {
 
 export const setNoteImportance = async (req: Request, res: Response) => {
     try {
-        const { notesId } = req.params;
+        const { noteId } = req.params;
         const { isImportant } = req.body;
         if (!isImportant) {
             return res.status(400).json({ "message": "Incomplete Information" });
         }
-        const note = await GetNoteById(notesId);
+        const note = await GetNoteById(noteId);
         if (!note) {
             return res.status(400).json({ "message": "Note Id not matched" });
+        }
+        const currentUserId = get(req, 'identity._id') as string | undefined;
+        if (!currentUserId) {
+            return res.sendStatus(403);
+        };
+        if (currentUserId.toString() !== note.userId) {
+            return res.status(400).json({ "message": "Ownership mismatch" })
         }
         note.isImportant = isImportant;
         await note.save();
@@ -85,8 +98,19 @@ export const setNoteImportance = async (req: Request, res: Response) => {
 export const deleteNote = async (req: Request, res: Response) => {
     try {
         const { noteId } = req.params;
-        const note = await DeleteNoteById(noteId);
-        return res.status(200).json(note).end();
+        const note = await GetNoteById(noteId);
+        if (!note) {
+            return res.status(400).json({ "message": "Note Id not matched" });
+        }
+        const currentUserId = get(req, 'identity._id') as string | undefined;
+        if (!currentUserId) {
+            return res.sendStatus(403);
+        };
+        if (currentUserId.toString() !== note.userId) {
+            return res.status(400).json({ "message": "Ownership mismatch" })
+        }
+        const deletedNote = await DeleteNoteById(noteId);
+        return res.status(200).json(deletedNote).end();
     } catch (err) {
         console.log(err);
         return res.sendStatus(400);
